pyproject.toml
name = "schedule"
version = "0.1.0"
description = ""
authors = ["Davlatbek Yuldashov <davlatyuldashov404@gmail.com>"]
readme = "README.md"
packages = [{include = "schedule_bot"}]

[tool.poetry.dependencies]
python = "3.11.4"
aiogram = "3.0.0rc1"
python-dotenv = "1.0.0"

[build.system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"